requirements.txt
requests==2.27.1
SQLAlchemy==1.4.32
retry2==0.9.3
Flask==2.0.3
